[BlockStorage] taking advantage of traits to reduce duplicate code
