+++ src/BlockStorage/Requests/DeleteBlockStorageRequest.php
@@ -2,6 +2,7 @@
 
 namespace wappr\DigitalOcean\BlockStorage\Requests;
 
+use wappr\DigitalOcean\BlockStorage\VolumeID;
 use wappr\DigitalOcean\Contracts\ModelInterface;
 use wappr\DigitalOcean\Contracts\BlockStorage\DeleteBlockStorageInterface;
 use wappr\DigitalOcean\Contracts\Requests\RequestModel;
@@ -11,26 +12,5 @@ use wappr\DigitalOcean\Contracts\Requests\RequestModel;
  */
 class DeleteBlockStorageRequest extends RequestModel implements ModelInterface, DeleteBlockStorageInterface
 {
-    /**
-     * @var string
-     */
-    protected $drive_id;
-
-    /**
-     * DeleteBlockStorageRequest constructor.
-     *
-     * @param string $drive_id
-     */
-    public function __construct(string $drive_id)
-    {
-        $this->drive_id = $drive_id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getDriveId(): string
-    {
-        return $this->drive_id;
-    }
+    use VolumeID;
 }
