+++ src/BlockStorage/Requests/RetrieveBlockStorageSnapshotsRequest.php
@@ -2,6 +2,7 @@
 
 namespace wappr\DigitalOcean\BlockStorage\Requests;
 
+use wappr\DigitalOcean\BlockStorage\VolumeID;
 use wappr\DigitalOcean\Contracts\ModelInterface;
 use wappr\DigitalOcean\Contracts\BlockStorage\RetrieveBlockStorageSnapshotsInterface;
 use wappr\DigitalOcean\Contracts\Requests\RequestModel;
@@ -11,26 +12,5 @@ use wappr\DigitalOcean\Contracts\Requests\RequestModel;
  */
 class RetrieveBlockStorageSnapshotsRequest extends RequestModel implements ModelInterface, RetrieveBlockStorageSnapshotsInterface
 {
-    /**
-     * @var string
-     */
-    protected $volume_id;
-
-    /**
-     * RetrieveBlockStorageSnapshotsInterface constructor.
-     *
-     * @param string $volume_id
-     */
-    public function __construct(string $volume_id)
-    {
-        $this->volume_id = $volume_id;
-    }
-
-    /**
-     * @return string
-     */
-    public function getVolumeId(): string
-    {
-        return $this->volume_id;
-    }
+    use VolumeID;
 }
